requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
dnspython==2.7.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonwebtoken==1.0.0
MarkupSafe==3.0.2
msgspec==0.19.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2022.7.1
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
